Moved subcommand certificaterequest to its own package
