+++ cmd/ctl/pkg/create/BUILD.bazel
@@ -2,27 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "certificaterequest.go",
-        "create.go",
-    ],
+    srcs = ["create.go"],
     importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/create",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/apis/certmanager/v1alpha2:go_default_library",
-        "//pkg/client/clientset/versioned:go_default_library",
-        "//pkg/util/pki:go_default_library",
-        "//pkg/webhook:go_default_library",
+        "//cmd/ctl/pkg/create/certificaterequest:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
         "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
-        "@io_k8s_cli_runtime//pkg/resource:go_default_library",
-        "@io_k8s_client_go//rest:go_default_library",
         "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
-        "@io_k8s_kubectl//pkg/util/i18n:go_default_library",
-        "@io_k8s_kubectl//pkg/util/templates:go_default_library",
     ],
 )
 
@@ -35,7 +22,10 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [":package-srcs"],
+    srcs = [
+        ":package-srcs",
+        "//cmd/ctl/pkg/create/certificaterequest:all-srcs",
+    ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
 )
