+++ cmd/ctl/pkg/create/create.go
@@ -20,6 +20,8 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
+
+	"github.com/jetstack/cert-manager/cmd/ctl/pkg/create/certificaterequest"
 )
 
 func NewCmdCreate(ioStreams genericclioptions.IOStreams, factory cmdutil.Factory) *cobra.Command {
@@ -29,7 +31,7 @@ func NewCmdCreate(ioStreams genericclioptions.IOStreams, factory cmdutil.Factory
 		Long:  `Create cert-manager resources e.g. a CertificateRequest`,
 	}
 
-	cmds.AddCommand(NewCmdCreateCR(ioStreams, factory))
+	cmds.AddCommand(certificaterequest.NewCmdCreateCR(ioStreams, factory))
 
 	return cmds
 }
